requirements.txt
pyserial==3.5
matplotlib==3.10.1
numpy==2.2.5
pandas==2.2.3
Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1

# Librerías para Machine Learning
scikit-learn==1.5.2
seaborn==0.13.2
tensorflow==2.18.0
xgboost==2.1.3
lightgbm==4.5.0
catboost==1.2.7